Some progress on test array reductions
This could work, but it'll take too long to fix up all the
implementations. Key points:

- Relying on a little NumPy for boxing things up again. This needs to be
  eventually replaced with returning a capsule with DLPack or something
- Using the array interface to access NumPy arrays without typing them.
  This will be slow, could replace with DLPack?
- The C side is still doing things with NumPy, that can be fixed later.

+++ numba/cpython/old_hashing.py
@@ -226,7 +226,7 @@ def _long_impl(val):
 def int_hash(val):
 
     _HASH_I64_MIN = -2 if sys.maxsize <= 2 ** 32 else -4
-    _SIGNED_MIN = types.int64(-0x8000000000000000)
+    _SIGNED_MIN = -0x8000000000000000
 
     # Find a suitable type to hold a "big" value, i.e. iinfo(ty).min/max
     # this is to ensure e.g. int32.min is handled ok as it's abs() is its value
